refactor(openjiuwen): extract MemoryRuntimeRail to its own file
Move the 143-line nested static class out of OpenJiuwenAgentRuntimeHandler
into a package-private top-level file. The handler shrinks from 423 to 252
lines. Visibility is unchanged (both files share the engine.openjiuwen
package). The rail calls OpenJiuwenAgentRuntimeHandler.errorMessage() which
is protected static and accessible within the package. Tests updated from
the qualified OpenJiuwenAgentRuntimeHandler.MemoryRuntimeRail reference to
the unqualified top-level name.
